chore(item): throw warning if edge src or target does not exist
